Hotspots might work now, physics needs testing

+++ src/high_level_calcs.hpp
@@ -687,13 +687,13 @@ public:
                     {
                         auto env_func_ = std::get<envelope_func>(env_func);
 
-                        std::poisson_distribution<uint_fast8_t> dist(sigma_jet*AA_params.Tpp->at(newpair));
-                        uint_fast8_t nof_dijets = dist(*eng);
+                        std::poisson_distribution<uint_fast16_t> dist(sigma_jet*AA_params.Tpp->at(newpair));
+                        uint_fast16_t nof_dijets = dist(*eng);
                         newpair.dijets.reserve(nof_dijets);
 
                         const double sqrt_s = newpair.getcr_sqrt_s();
 
-                        for (uint_fast8_t i=0; i < nof_dijets; i++)
+                        for (uint_fast16_t i=0; i < nof_dijets; i++)
                         {
                             auto dummy = pqcd::nn_coll_params
                                 (
@@ -883,7 +883,7 @@ public:
                     {
                         auto env_func_ = std::get<envelope_func>(env_func);
 
-                        std::poisson_distribution<uint_fast8_t> dist(sigma_jet*AA_params.Tpp->at(new_pair));
+                        std::poisson_distribution<uint_fast16_t> dist(sigma_jet*AA_params.Tpp->at(new_pair));
                         uint_fast16_t nof_dijets = dist(*eng);
                         new_pair.dijets.reserve(nof_dijets);
 
@@ -934,13 +934,13 @@ public:
                     {
                         auto env_func_ = std::get<envelope_func>(env_func);
 
-                        std::poisson_distribution<uint_fast8_t> dist(sigma_jet*AA_params.Tpp->at(new_pair));
-                        uint_fast8_t nof_dijets = dist(*eng);
+                        std::poisson_distribution<uint_fast16_t> dist(sigma_jet*AA_params.Tpp->at(new_pair));
+                        uint_fast16_t nof_dijets = dist(*eng);
                         new_pair.dijets.reserve(nof_dijets);
 
                         const double sqrt_s = new_pair.getcr_sqrt_s();
 
-                        for (uint_fast8_t i=0; i < nof_dijets; i++)
+                        for (uint_fast16_t i=0; i < nof_dijets; i++)
                         {
                             auto dummy = pqcd::nn_coll_params
                                 (
